+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-08-03 22:51:25
-// Commit: chore(theme): update dependencies and cleanup unused imports in theme
+// Updated at: 2026-08-04 13:33:56
+// Commit: style(theme): harmonize light and dark theme palette for theme
 
 class themeComponentState {
-  final String commitRef = '3336';
-  final String lastMessage = 'chore(theme): update dependencies and cleanup unused imports in theme';
-  final DateTime timestamp = DateTime.parse('2026-08-03T22:51:25+05:00');
+  final String commitRef = '3347';
+  final String lastMessage = 'style(theme): harmonize light and dark theme palette for theme';
+  final DateTime timestamp = DateTime.parse('2026-08-04T13:33:56+05:00');
 }